+++ src/main/java/MainContainer.java
@@ -10,19 +10,8 @@ public class MainContainer {
     private static final Logger log = LogManager.getLogger();
     public static final int PROD_PORT = 443;
     public static final int TEST_PORT = 8080;
-    public static void main(String[] args) {
 
-        /*
-        Properties properties = new Properties();
-        try {
-            //properties.load(MainContainer.class.getClassLoader().getResourceAsStream("ibs.properties"));
-            //ConfigValueHandler.PRODUCT_VERSION.activate(properties.getProperty(ConfigValueHandler.PRODUCT_VERSION.systemName));
-            //ConfigValueHandler.PRODUCT_VERSION.activate(MainContainer.class.getPackage().getImplementationVersion());
-        } catch (IOException e) {
-            log.error("error");
-            throw new RuntimeException(e);
-        }
-        */
+    public static void main(String[] args) {
 
         if (args.length == 0) {
             log.info("In Prod Mode - SSL");