+++ src/main/java/com/adobe/campaign/tests/service/JavaCalls.java
@@ -1,19 +1,23 @@
 package com.adobe.campaign.tests.service;
 
+import com.adobe.campaign.tests.service.exceptions.IBSConfigurationException;
+import com.adobe.campaign.tests.service.exceptions.IBSRunTimeException;
+import com.adobe.campaign.tests.service.exceptions.NonExistantJavaObjectException;
+
 import java.util.*;
+import java.util.stream.Collectors;
 
 public class JavaCalls {
 
     private Map<String, CallContent> callContent;
 
-   // private Map<String,String> environmentVariables;
-   private Map<String,Object> environmentVariables;
+    private Properties environmentVariables;
 
     private IntegroBridgeClassLoader localClassLoader;
 
     public JavaCalls() {
         callContent = new LinkedHashMap<>();
-        environmentVariables = new HashMap<>();
+        environmentVariables = new Properties();
         setLocalClassLoader(new IntegroBridgeClassLoader());
     }
 
@@ -63,22 +67,31 @@ public class JavaCalls {
     }
 
     private void updateEnvironmentVariables() {
-        CallContent l_setEnvironmetVars = new CallContent();
-        l_setEnvironmetVars.setClassName(ConfigValueHandler.ENVIRONMENT_VARS_SETTER_CLASS.fetchValue());
-        l_setEnvironmetVars.setMethodName(ConfigValueHandler.ENVIRONMENT_VARS_SETTER_METHOD.fetchValue());
+        CallContent l_setEnvironmentVars = new CallContent();
+        l_setEnvironmentVars.setClassName(ConfigValueHandler.ENVIRONMENT_VARS_SETTER_CLASS.fetchValue());
+        l_setEnvironmentVars.setMethodName(ConfigValueHandler.ENVIRONMENT_VARS_SETTER_METHOD.fetchValue());
+
+        List<Object> badVariables = getEnvironmentVariables().keySet().stream().filter(k -> getEnvironmentVariables().getProperty((String) k)==null).collect(
+                Collectors.toList());
+
+        if (badVariables.size()>0) {
+            throw new IBSRunTimeException("The given environment variables should only contain strings.\n"+badVariables);
+        }
 
         //Fetch all environment variables
-        Properties argumentProps = new Properties();
-        environmentVariables.keySet().forEach(k -> argumentProps.put(k, environmentVariables.get(k)));
-        l_setEnvironmetVars.setArgs(new Object[] { argumentProps });
-        l_setEnvironmetVars.call(this.getLocalClassLoader());
+        l_setEnvironmentVars.setArgs(new Object[] { environmentVariables });
+        try {
+            l_setEnvironmentVars.call(this.getLocalClassLoader());
+        } catch (NonExistantJavaObjectException nejoe) {
+            throw new IBSConfigurationException("The given environment value handler "+ConfigValueHandler.ENVIRONMENT_VARS_SETTER_CLASS.fetchValue()+ "."+ConfigValueHandler.ENVIRONMENT_VARS_SETTER_METHOD.fetchValue()+ " could not be found.", nejoe);
+        }
     }
 
-    public Map<String, Object> getEnvironmentVariables() {
+    public Properties getEnvironmentVariables() {
         return environmentVariables;
     }
 
-    public void setEnvironmentVariables(Map<String, Object> environmentVariables) {
+    public void setEnvironmentVariables(Properties environmentVariables) {
         this.environmentVariables = environmentVariables;
     }