+++ src/main/java/com/adobe/campaign/tests/service/exceptions/IBSConfigurationException.java
@@ -1,0 +1,13 @@
+package com.adobe.campaign.tests.service.exceptions;
+
+/**
+ * This exception is intended to be used for issues related to configuration of the IBS.
+ *
+ * author: gandomi
+ */
+public class IBSConfigurationException extends RuntimeException {
+
+    public IBSConfigurationException(String message, NonExistantJavaObjectException nono) {
+        super(message,nono);
+    }
+}